[#223] Add initial tariff plans, fix setting of default plan, fix migrations, fix tariff plan views and styles

+++ db/migrate/20131025055523_add_tariff_to_company.rb
@@ -3,6 +3,6 @@ class AddTariffToCompany < ActiveRecord::Migration
     add_column :companies, :tariff_id, :integer
     add_column :companies, :user_payment_id, :integer
     add_column :companies, :tariff_end, :datetime
-    add_column :companies, :paid, :float
+    add_column :companies, :paid, :decimal
   end
 end
